+++ spec/shipcloud/shipment_spec.rb
@@ -98,6 +98,13 @@ describe Shipcloud::Shipment do
         and_return(true)
       Shipcloud::Shipment.delete("123")
     end
+
+    it "doesn't raise an error" do
+      stub_request(:delete, "https://your-api-key:@api.shipcloud.io/v1/shipments/123").
+        to_return(status: 204, body: "")
+      expect { Shipcloud::Shipment.delete("123", api_key: "your-api-key") }.
+        to_not raise_error
+    end
   end
 
   describe ".all" do